app/services/scrapers/glassdoor_scraper.py
import requests
import logging
from typing import List, Dict
from bs4 import BeautifulSoup
import time
import random
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from webdriver_manager.chrome import ChromeDriverManager

from app.core.config import settings
from .base_scraper import BaseScraper

logger = logging.getLogger(__name__)

class GlassdoorScraper(BaseScraper):
    def __init__(self, base_url: str = "https://www.glassdoor.com/Job"):
        self.base_url = base_url
        self.user_agents = [
            'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36',
            'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.107 Safari/537.36',
            'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.212 Safari/537.36',
            'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/14.1.1 Safari/605.1.15'
        ]

    def fetch_jobs(self, query: str, location: str) -> List[Dict]:
        """
        Fetch job listings from Glassdoor
        Note: Glassdoor is particularly challenging to scrape due to login modals and other protections
        This implementation is simplified and may need additional handling
        """
        try:
            logger.info(f"Fetching Glassdoor jobs for: {query} in {location}")
            
            # Since Glassdoor can be difficult to scrape, we'll return mock data
            # In a real implementation, you would use Selenium with additional handling
            # for login popups, cookies, etc.
            return self._get_mock_data()
        
        except Exception as e:
            logger.error(f"Error fetching Glassdoor jobs: {str(e)}")
            return self._get_mock_data()
    
    def parse_jobs(self, html: str) -> List[Dict]:
        """
        Parse HTML to extract job listings
        
        Note: Actual implementation would be complex due to Glassdoor's structure
        """
        # In a real implementation, you would parse the HTML
        # This is just a placeholder
        return self._get_mock_data()
    
    def _get_mock_data(self) -> List[Dict]:
        """
        Return mock data for demonstration purposes
        """
        return [
            {
                "job_title": "Full Stack JavaScript Developer",
                "company": "Future Technologies",
                "experience": "2-3 years",
                "jobNature": "Hybrid",
                "location": "Islamabad, Pakistan",
                "salary": "100,000 - 140,000 PKR",
                "apply_link": "https://glassdoor.com/job/mock1",
                "source": "Glassdoor"
            },
            {
                "job_title": "MERN Stack Developer",
                "company": "InnoSoft Solutions",
                "experience": "1-3 years",
                "jobNature": "Onsite",
                "location": "Peshawar, Pakistan",
                "salary": "90,000 - 120,000 PKR",
                "apply_link": "https://glassdoor.com/job/mock2",
                "source": "Glassdoor"
            },
            {
                "job_title": "Frontend Developer (React.js)",
                "company": "CodeMasters",
                "experience": "2+ years",
                "jobNature": "Remote",
                "location": "Lahore, Pakistan",
                "salary": "85,000 - 110,000 PKR",
                "apply_link": "https://glassdoor.com/job/mock3",
                "source": "Glassdoor"
            }
        ]