requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
httpx==0.24.0
beautifulsoup4==4.12.0
selenium==4.8.3
webdriver-manager==3.8.5
python-dotenv==1.0.0
openai==0.27.4
pytest==7.3.1
pytest-asyncio==0.21.0
lxml==4.9.2
passlib==1.7.4